+++ release-notes.md
@@ -2,6 +2,7 @@
 * 147110 - Build: Readiness to open meeting (ROM)
 * 179195 - Build: Show only statuses relevant to Presumption route
 * 147094 - Build: Pre-funding agreement checkpoint meeting (Central route)
+* 178584 - Build: Principal designate (both routes)
 
 ## 0.12.0
 * 175400 - Build: Update 'You do not have access to the service' page with access form link